Reduce the commDiag.chpl regression to 1 get per remote locale vs. 3
Go onto the locale that owns the barrier instance in the wrapper procedures
instead of in the inner procedures.  This reduces the number of remote gets
without increasing the number of forks.

Updated the commDiag.chpl test and comm-gasnet.good file to make the output
easier to read.

There is still one unexpected get per locale (except the one that owns the
barrier object) per coforall+barrier kernel.  It is coming from an endcount
object for the coforalls.