Rearrange the components of an internal error
As others have commented, the logic around printing out errors and
warnings is completely spaghetti.  I thought that reordering the
elements of the internal error messages would be trivial, but it's
not and has taken me awhile to reach this point, so I'm pausing
to catch my breath before going on.