Update testReleasesPerformance for new start_test
Code-wise, there wasn't anything required here, I just noted that you
have to set your venv location for the test system before launching
it (since, otherwise, it will try to use venv environments for the
old releases which didn't have/support them).

While here, I also noted that if we did a --gen-graphs at the end of
the run, the results could be inspected and verified more easily, so
did that.