Fix annotation generation
Our annotations require PyYAML. This used to be installed on our testing
machines, but doesn't appear to be anymore. To get around this I'm just adding
it to the list of test-venv requirements for performance runs. We don't have
approval to add it to the test-venv in general, so I'm just appending it in
common-perf