Add tests for layout equality
These tests test for equality between different layouts.  For
default layouts, we use a single domain map for default layouts
of rectangular, associative, sparse, and opaque domain maps
unfortunately, so cannot distinguish between these domain map
types.  Added a future suggesting that we split DefaultDist
into a default dist per major domain type to address this
(layoutEquality.chpl)

Added the equality dsi function to the CSR layout so that it
can be distinguished from a default sparse layout and a test
to lock this in (layoutEquality2.chpl).