Free heap allocated variables used in cobegin and coforall statements
This will resolve all the memory leaks caused by enabling parallel array
initialization with fce777eb4224b84b17e0a965d5f2915fd130279d It also reduces
the exiting memory leaks for memory/sungeun/refCount/domainMaps.

Freeing heap allocated variables was originally added with
04649adee918d0a15dc4a8f365ba83c51831c2cc and
cf2a0a0c1d4474d5ca9ca44d5cddddde20d6d3f7. However, this code was careful to not
free heap allocated variables if they could be used by a task function (begin,
cobegin, coforall)

This was to avoid freeing heap allocated variables that could still be in use
by an asynchronous task. Prior to task intents, we would have had a really hard
time detecting if a heap allocated variable was used inside a begin that was
inside another task function. Because of this, we conservatively assumed that
any heap allocated var passed to any task function could then be passed to a
begin, so we didn't free them.

With task intents, variables used in a begin that outlive the scope they were
declared in will be passed by ref to the begin and any enclosing task
functions. Because of this we no longer need to check for cobegin/coforall
statements as we're be able to directly see if heap vars are used in a begin.

Note that this commit removes the check for FLAG_COBEGIN_OR_COFORALL and
FLAG_NON_BLOCKING. When the code to free heap variables was originally added
begin+on and coforall+on functions weren't always marked with FLAG_BEGIN or
FLAG_COBEGIN_OR_COFORALL so we also had to check the FLAG_NON_BLOCKING. Since
then the AST has been normalized and we don't have to worry about checking for
FLAG_NON_BLOCKING anymore.

Long term I think we want to free heap allocated variables used in begin
statements too. We're moving towards not automatically keeping variables used
by asynchronous tasks alive passed their lexical scope. Making this change
resulted in ~5 regressions, all of which rely on such behavior. That seems like
a big change right before code freeze so that can wait until after the release.

As background for the memory leaks:

The problem was that we leaked 16 bytes (a wide pointer to a DefaultRectangular
array) when we did a DefaultRectangular dsiReindex, dsiSlice, or dsiRankChange
with gasnet segment fast/large.

With gasnet segment fast/large, we heap allocate all arguments passed to on
functions (see parallel.cpp:needHeapVars)

The above functions all declare an 'alias' variable, then inside an on
statement they update 'alias', then after the on statement they return 'alias'.
This code causes 'alias' to be passed as an argument to the on function so it
gets heap-ified for gasnet fast/large.

This alias has always been heap allocated for gasnet fast/large. What changed
with the parallel init commit is that it's no longer being freed. We have code
in parallel.cpp to free heap allocated variables, however, we do not free
variables that might be used in a task function. Parallel init added a coforall
to the DefaultRectangular initialization code, so we're no longer freeing the
wide pointer to 'alias'.

Another way to resolve these regressions would be to stop heap allocating on
function arguments for gasnet fast/large and rely on out of segment gets/puts
like we do for ugni. I tried this, but every test that actually uses two or
more locales segfaulted at runtime. I believe that our our support for gasnet
out of segment gets/puts needs more work. This is not code that I'm familiar
with yet and it seems like it would be a big change this close to code freeze.