Remove numa specific .good files for memory/figueroa/LeakedMemory[2, 5, 6]
With daf0a0de605fb63e91b18ab5a8bb1dd63b52bb98 we free more variables allocated
on the heap. This reduced numa leaks for memory/figueroa/LeakedMemory[2, 5, 6]
to the point that the leaks are the same as every other configuration. This
means that the numa specific .good files are no longer needed.