Add PyYAML to our testing system requirements
Our annotations rely on PyYAML, so I'm adding it to the list of requirements.
Previously our perf testing machines happened to have PyYAML installed so we
didn't think to have it in our requirements. This will allow perf graph
annotations to be generated on any machine.

Since PyYAML should always be available, I added a warning to genGraphs if we
fail to import our annotate module (which would only happen if PyYAML isn't
available.)