Allow precomps to put all their output in the log file
I discovered that in the case of error messages in a precomp, the output when
running would be different from the log file generated.  Basically, the error
messages were not getting stored when every other output was.  Elliot posited
that this was also true for prediffs and preexecs, so my next step is to get
those working.  This commit only handles precomps, though.