Fixing a segfault in the glob() iterator when it's a follower
[reviewed and code reviewed by @ronawho]

Due to an off-by-one error, the glob() iterator would sometimes
segfault when asked to be follow the range 0..0.  Specifically,
I was forgetting to take the 0-shifting of the followThis range
into account and so would think that it was OK to follow 0..0
(a 1-element range) when there were 0 matches on the glob() call.
This would then cause us to request the initial match, but there
weren't any, so a segfault would occur.  This would happen
nondeterministically when the number of tasks was greater than
10 when following the range 1..10 depending on whether the task
that owned 0..0 executed before or after another task that
would complain that there were not enough iterations.

While here, I updated the error message slightly to give more
information.

This is a second attempt at this issue -- the first was at
PR #2516