Comment out the return type for string.localize()
Specifying the return type vs leaving it implicit still causes
significant code differences. In the case of `string.localize()` it will
cause an op= into the return variable, which causes us to always copy
the string. This leads to very poor performance on some tests since the
localize call is assumed to be cheap when the data is already local.