Quiet parallel/forall/vass/memleaks* regression
I had to run ~1000 times to see the regression so it appears to be very
sporadic and we were unlucky enough for it to occur on the first night of
testing.

This adds a serial to the forall statements to quiet the regressions. Serial
was attached to the forall in memleaks2-BlockDist, but I thought I could remove
it in 880d1520e57c8b9979d4ea9fce47e48dc520f22e. I'm re-adding that and adding
it to memleaks1-minimal as well.

See JIRA issue 112: https://chapel.atlassian.net/browse/CHAPEL-112