Add halt to spawn() for CHPL_COMM=ugni
fork doesn't currently work with ugni when memory is registered with the NIC
and it will result in a segfault. Until we resolve that issue, this adds a halt
and nice error message to tell users what went wrong so they aren't just hit
with an unexplained segfault.

Note that the runtime spawn code will use vfork over fork when stdin, stdout,
stderr=FORWARD so we don't halt in that case. Additionally memory isn't
registered with the NIC when numLocales == 1, so we don't halt then either.

See JIRA issue 113 for more info: chapel.atlassian.net/browse/CHAPEL-113