Add a test to demonstrate the limitations of Spawn under ugni
Under ugni, spawn doesn't work if numLocales > 1 and one of stdin, stdout,
stderr != FORWARD. This adds a 2-locale copy of spawn-system that runs with
stdout=FORWARD and stdout=PIPE to demonstrate that fork will work for a
multi-locale program with stdout=FORWARD and that our error message about when
Spawn will fail is triggered for stdout=PIPE

The regular spawn tests already show that fork works with numLocales == 1.

Note that this is only run with CHPL_COMM=ugni. That does mean it won't
actually get run nightly, but this is mostly meant to help us resolve not being
able to fork under ugni. See JIRA issue 113 for more details:
chapel.atlassian.net/browse/CHAPEL-113