In the spec, reconcile pdf page numbers
The observable page numbers in the spec used to be off by 1
from the printed page numbers, starting with the Contents page.
It was the third page in the document, while being labeled as Page 2.

This commit makes the Contents page be the second page,
still labeled as such.

While there -- I also moved "(C) 2015" into the title page,
eliminating the almost-empty second page in the document.

I am not sure that the new place of the copyright is great.
Since we are using the \maketitle command for that page,
there is not much flexibility. We could put copyright at the bottom
using \thanks within either \author or \title. That will not be
as pretty - it will be a left-justified footnote.
Still, having an almost-empty page with just the copyright
at the bottom does not feel worth the aestetic advantage to me.