Content update, mainly guard pages for {gasnet,none}+muxed.
The biggest change here is a discussion of the use of guard pages in
muxed tasking.  There are also some smaller clarifications and edits.
Some of these led to my reformatting whole paragraphs, making it look
like the changes are bigger than they really are.  I also fixed a few
header underlines, because I think the hierarchy was wrong in a couple
of spots.