For distributions, given it is code freeze so we do not get to modify 'chpldoc', I took the approach - suggested by Elliot - to put all documentation into the comment associated with the distribution class decl.
fixDistDocs.perl ensures that no other decls for each file
make it into the html page.

I ended up with a style of presenting each distribution that's somewhere
between a formal spec and a narrative-like tutorial. We can revisit that.

The diffs for ReplicatedDist.chpl show a couple of param decls moved.
That's just because of 'diff' - there are no code changes,
at least not intentional.

I listed some chpldoc TODOs in each file. Most notably:

    Move the notes about constructors and selected other functions
    from that monolithic comment to the comment on the actual declaration.

    There will be a challenge because some constructors use the functions
    like _determineRankFromStartIdx() and _determineIdxTypeFromStartIdx()
    that we might not want the users to know about.

    The benefit of such a move is that the documentation will be kept
    in sync with the code.

    I rephrased and toned down this piece from the spec - OK?

    It is parameterized by the rank and index type of the domains it
    supports. Thus domains of different ranks or different index types
    must be distributed with different Cyclic [or whatever] distributions.

    Need automatic testing of examples that appear in chpldocs.

Separately, we have been considering chpldocs for Tuples to replace
the section "Predefined Functions and Methods on Tuples" in the spec.
For now we decided to keep that section and at the same time
provide chpldocs for those functions.