ReStructuredText-ify modules/internal/ChapelLocale.chpl, et al.
Because chpldocing ChapelLocale.chpl ends up bringing it into the early
stages of the compiler twice and thus causes errors, this chpldocing is
instead done in the new file ChapelLocale-chpldoc.chpl, which the
compiler doesn't bring in automatically.  As a nice side effect, this
also allows us to put the definitions of callStackSize, numLocales,
LocaleSpace, and Locales directly in this file and document them
properly, rather than either documenting them in the widely separated
places where they are really defined, or documenting them in the header
comment for this file.

After the release we plan to clean all this up.