Chpldoc vectorizeOnly iterator
Chpldoc the vectorizeOnly iterator which currently lives in
ChapelIteratorSupport. This mostly just adds a top level comment for the
module, and cleans up the comment that was there. It also adds a few no-docs
and renames the module.

For next release, this should be moved to it's own standard module, and I
should add a vectorization primer.