Quiet forallInStandalone regression
This test is part of the vectorization suite that reports line numbers for user
level loops that generate vectorization hints. However, for standalone
iterators, we currently report the line number of the yield points in the
standalone iterator. Documentation was added to ChapelRange so the reported
line number changed. I'm just updating the reported line number here.

Right after the release, I'll convert this test to a future so it doesn't cause
noise because of changes to our ChapelRange line numbers. I'll also take a look
and see if it's easy enough to just fix whatever is causing us to not report
user line numbers.