Removed Standard Modules chapter in spec
It has been replaced with online docs.

Brad and Vass did not feel strongly about having a reference
to online docs in the spec. So to save time for now we will go
without such a reference.

The Standard Modules chapter discussed the difference between
"automatic" and "optional" modules, which we are not preserving.
Instead we want the online module docs to indicate which things
are available automatically. The other modules would implicitly
require to be "use"-ed.