Merge pull request #2717 from lydia-duncan/updateTechnoteReferences
Update references to doc/technotes/* [reviewed by @ben-albrecht]

There were a couple of mentions of these in the modules Curl and Help, the
third-party and llvm READMEs and the man page. I tested the man page and uses
of the modules mentioned in our testing system and observed no problems.

I didn't update HDFS.chpl since I was told @gbtitus already handled it.