Remove last remaining use of pragma "export" and turn off pragma form
The test variables/bradc/pragmaMultipleVars.chpl was intended to ensure that
a pragma applied to a multiple variable declaration would affect all variables
in the declaration.  This is independent of the "export" pragma, so can safely
be changed into a pragma we expect to stick around longer.  I arbitrarily chose
"no doc" as its replacement; the other good option that occurred to me was
"locale private" but I figured that pragma only made sense in a multi-locale
context, and "no doc", while limited to chpldoc runs, was always visible after
parse time in that context and we test chpldoc in most places.  The prediff and
skipif settings were no longer necessary, though some additional files were
needed to test with chpldoc.

Since this is the last use of the pragma form of export, it can be removed and
left with only the FLAG form.