Fix `make check` in nightly (round 2)
#2728 fixed `make check` in nightly by source-ing a setchplenv script first.
However, it was still failing on ubutnu based systems. This was because perl's
system() function uses /bin/sh and not /bin/bash. In bash `.` and `source` are
synonymous, but that's not necessarily true for sh.  Since it's /bin/sh, the
`.` operator is the right thing to use instead of `source`