Include a locale field in string record
We've generally been trying to notify a string if it
is copied across locales (by calling some function
in String.chpl), but that has been challenging. This
is an attempt to understand how far the alternative
gets us.

The alternative is to store the locale of each string
within that string, so that the string data can be
silently copied across locales and everything will
still work (since the locale will now point to the
place really storing the string data). I did that
here by adding a locale_id field to the string record;
an alternative would be to make the c_ptr buffer
be instead a wide pointer.

One issue is that it does not appear that we can use
the locale type itself within String.chpl due to internal
module ordering issues. So, instead, we use chpl_nodeID_t
and explicitly use the chpl_on_locale_num primitive in
on statements. Both of these could be improved upon.