Stop heap-allocating simple stack variables
Historically, we've heap-allocated variables that are logically "stack"
variables in the event that they might outlast their lexical scope due
to being referred to within a begin task.  We've recently decided to
stop doing this, but haven't cleaned up    places where we        were working
too hard to preserve the old semantics.     This looks like low-hanging
fruit as determined by observing the generated code for this program:

var flag$, flag2$: sync bool;
{
  var xyz: int;

  begin with(ref xyz) {
    xyz = 123;
    writeln("xyz is: ", xyz);
    flag2$ = true;
    flag$;  // block until flag is set
    writeln("xyz is: ", xyz);
  }
  flag2$;  // block until flag2 is set
  writeln("xyz is: ", xyz);
}
flag$ = true;

Prior to this change, 'xyz' was stored as a heap_int64_t, whereas
now it's a simple int64_t.