Add explicit ref intent for 'this'
This applies a part of 9ee323f
which I reverted on string-as-rec in PR #2729
plus a couple of similar cases.

E.g.:

  proc ~regexp() --> proc ref ~regexp()

This change is needed because the destructor updates one of the fields
of 'this'. Without the explicit 'ref' intent, 'this' implicitly has
'const ref' concrete intent, so it is illegal to update its fields.
The compiler lets this code pass because it lacks checking for this case.