Really fix this test
Oops, previous fix actually caused a (potential?  definite?) deadlock.
Noting that 'b' does not need to be passed in by reference to the
'begin' task because it isn't actually used simultaneously by anyone
else, I'm changing it to an 'in' intent which is sufficient for this
program and removes the race that the task may refer to 'b' after it
leaves scope.