Revert "Stop heap allocating because of non-blocking ons"
#2745 caused some ugni-qthreads regressions, so we're going to revert it for
the time being. The issue is basically that the target variable of a network
AMO has to registered with the NIC (and can't be trampolined through a
registered segment.)

Under ugni-muxed, task stacks are allocated out of the registered heap, but
qthreads just grabs task stacks from the system allocator. #2475 stopped heap
allocating some arguments to nonblocking on functions causing regressions for
SSCA2 and a few performance tests.

We either need to force heap allocation for network AMOs when stacks aren't
from registered heap memory, or update qthreads to grab stacks from the
registered heap. Either way, this isn't likely to be fixed in the next few
days, so we're reverting the change.

Note that we believe that the same issue probably exists for blocking on
functions, it's just not exposed by any tests. Whatever future work we do needs
to adds tests for these other cases