Update PRIM_WIDE_GET_ADDR/_wide_get_addr to return a UInt64
Storing an address in a signed int doesn't really make much
sense. It makes a little bit more sense to store an address in
an unsigned integer. Since _wide_get_addr right now is only
used for debugging, this change makes sense.

Since it is used for debugging, this code is not actually
run in any tests right now on master.

If we were to start using PRIM_WIDE_GET_ADDR elsewhere, it
might make sense to have it return some sort of pointer type
(e.g. c_void_ptr, if that is known to the compiler).

This change started out on string-as-rec as commit
bc18e40329f3fa5a9733445f1819113dc7ee2aec
but that commit included changes to other files, so I
opted to manually re-create a minimal patch.