Undo commenting out of halt() in #2734
That PR unintentionally renamed .locale.id to .locale_id on a c_string.
That was in _cast() from c_string to string. That generated a compile-time
error. Due to poor error handling in the compiler (mea culpa), the compiler
was not clear about the cause. To avoid the error, some code got commented out.

This corrects the typo and uncomments the code, reverting it to pre-2734.