Resurrection of a "pass syncs to generics" patch I developed under SVN
Vass's future from #2730 was something I tried to fix a few years
back, but failed to, in part because of how I/O relied on generics
w.r.t. syncs/singles.  However, I think Vass has improved the
situation since then, and this patch seems to make his future
pass, so I'm resurrecting it in hopes that it will be closer to
working today.