Repair --llvm after PR #2743
PR #2743 changed some code to not code generate the first argument of on
block functions to code to not generate arguments marked with
FLAG_NO_CODEGEN.

At the time, I failed to update the LLVM code generation code for those
cases. This patch updates the LLVM code generator in the corresponding
manner.