Back out change to flattenFunctions passByRef
The test test/types/records/ferguson/parallel/coforall.chpl
(a string-as-rec test) no longer functions if this code is
enabled. The pattern in that test is

  iter myiter() { ... yields non-pod record }
  coforall r in myiter() { ... use r }

If non-POD record types are always passed by reference in
flattenFunctions, by code generation we have a heap-allocated variable
(presumably storing iterator state), and the argument to the task function
becomes a reference to it, but that heap-allocated variable is freed
before the task runs.