Change calls from new_StringSymbol to new_CStringSymbol
These two functions are the same currently, but are different on
string-as-rec. Many of the places that have been converted are things
like prim_get_member calls, which take a string literal for the field
name to access. These seem like they should take some `IdentiferSymbol`
type instead in the long run.