Factor out some string handling code into unescapeString
In a few places in the compiler we want to handle a string with the raw
bytes rather than its escaped form ("\n" as one byte instead of two).
This change pulls some code out into a function that lets us perform
this transformation.

unescapeString is then used to fix a couple of bugs:

  string.length was incorrect when called on literals with escaped
  characters:
    "foo\nbar".length was 8, rather than 7.

  ascii(string) did not work for escaped characters:
    ascii("\n") would be 0x5C (\) rather than 0xA (newline)