On string-as-rec, revert #2439
#2439 on string-as-rec removed a couple of checks
in replaceVarUsesWithFormals().

This followed up our experiment on master, where we removed them
in #2362, observed it cause a valgrind regression, then reinstated in #2436.

We thought it would be good to have it on SAR, however right now we
believe it should behave the same way as it does on master,
so we are putting them back in on SAR.

To follow up: see if removing these checks (like in #2439)
still causes a valgrind regression in:

    domains/sungeun/sparse/index_not_in_domain_1.chpl

If it does, leave those checks in, like with this PR.
If it doesn't, take them out?