On string-as-rec, revert #2338
#2338 fixed an issue that started to arise due to #2197.
Given that #2197 was reverted in #2812, we do not need this fix,
so I am taking it out.

While the fix in #2338 is not needed at the moment, we should keep it in mind
in case we start getting more ref/deref temps, if that ever happens.

I am keeping the tests that #2338 introduced. They pass now,
as do the FileSystem tests that #2338 fixed