Add a new timer type to LCALS
Add a new timer type that uses the clock() function from time.h.  This matches
the default timing functionality in the official LCALS version.

The timer type can be chosen on the command line with the
config const defaultTimerType={Chapel|Clock|Cycle}.  "Chapel" is currently
the default and uses the Chapel standard "Time" module.  "Clock" uses the
clock() function from time.h.  "Cycle" is not currently implemented and
halts with a "not implemented" message.