Do not use calls to = in normalize for declared return types
If the return type is not declared, normalize generates code that uses
PRIM_MOVE to set the return value variable.

If the return type is declared however, before this patch, normalize would
generate a call to the = operator, starting with a "no-init" initialized
left-hand side.

We would like declared return types to have the return value PRIM_MOVE'd
into the return variable. The issue is that if there is a declared return
type, we want casts to get added if the types are coercible, and an error
to be generated if not. That is really what = is doing in this case.

However, normalize cannot detect the types being the same or different
since the types are not known until function resolution.

This patch takes the approach of adding a new primitive, PRIM_COERCE, that is
added in normalize for such cases. PRIM_COERCE is handled by function
resolution. PRIM_COERCE takes in two arguments: the rhs value and also the type
that we shoudl be coercing to (in this case, the function return type). When
the types are the same, the PRIM_COERCE is simply removed. Otherwise, re-create
the pattern of no-init/= that normalize used to generate. We could have used a
call to _cast in this case, but no-init/= is the pattern that works for
cases like this:

    config const n = 10;
    proc testproc(): [1..n] int {
      return for i in 1..n do i;
    }
    var A = testproc();
    writeln(A);

Besides replacing a call to "=" with PRIM_MOVE/PRIM_COERCE, this patch:

- adds a check for FLAG_TEMP before giving used-before-defined errors
  in normalize. I was getting errors with the return symbol not
  being defined. There is a comment indicating that this checking
  should perhaps not be done in normalize anyway.
- updated normalize_returns to avoid using PRIM_NO_INIT ever.
  Instead, the AST generated always uses PRIM_MOVE to set the
  return value.
- removes the warning in checkParsed for declared return types on
  param functions, since these now work after this patch.