Default to qthreads+hwloc for KNC instead of fifo
When we originally added support for KNC we saw poor performance with qthreads.
However, this was before we did any performance tuning with qthreads. In
particular we were using a work stealing scheduler, which did not perform well
especially as more cores were added.

Motivated by some recent work with stream and hwloc I decided to see if the
performance with qthreads is better now. The short summary is that
qthreads+hwloc performance is better, at least for stream (121 vs 115 GB/s.) As
a next step I'd like to add some nightly performance testing for KNC.

In the past hwloc didn't build on stream, but it appears to build just fine
now. This patch adds code to throw `-mmic` for KNC builds in the hwloc Makefile
just like we do for our other third-party libraries.

The specific changes in this patch are:
 - make qthreads and hwloc the defaults in the chplenv scripts
 - throw `-mmic` for KNC builds in hwloc Makefile
 - update documentation that mentioned fifo as the tasking layer for KNC