Don't refer to the non-stack atomic vars directly.
Instead, refer to them via ref arguments to the testing procs.  This
allows using the compile-time configuration to control which types we
declare and test.  Notably, we only do real(64) with CHPL_COMM=ugni.
(The code to do that one and to skip real(32) entirely was already
present, but I hadn't updated the integer types to be treated the same
way.)