Replace filenames in arg bundles with an index to a table
Filenames (added by insertLineNumbers) in arg bundles were being added
straight to the bundle, even though they are c_strings. Since c_strings
are only good on their original locales, use of the filenames passed into
the arg bundles would lead to segfaults. After this patch we now use the
string to find an index into a lookup table of filenames, pass the index
in the arg bundle, and use that index to get a valid c_string on the
other side.

This error would not show up on linux due to its weak ASLR, but would
cause major problems on OSX.