Add expiring value optimization
One common pattern is that a function creates some kind of record and
then that record is assigned to another variable.

In particular, consider a record R that manages memory (like a string
might). Let's think about using a function returnsR():R. There are two
cases:

 // initialization case
 var x = returnsR();

 // = assignment case
 var y:R;
 y = returnsR();

It is often the case that the result of a function call will be stored
in a compiler-introduced temporary. So we might have:

  // initialization case
  move tmp, returnsR()
  move x, initCopy(tmp)

  // = assignment case
  move tmp, returnsR()
  call = (x, tmp)

Both of these will create 1 more allocation than necessary.
They could be written this way with fewer allocations:

  // initialization case
  move x, returnsR()

  // = assignment case
  destroy x
  move x, returnsR()

This is a problem that move constructors solve in C++. Chapel has
less strict record semantics and so the move constructor business
can be implemented as an optimization.

This patch implements that optimization. The key limit for whether
or not the optimization can apply is if the right-hand-side value
in an initCopy or = call is dead at the time of the call.