Fix test_local_int_begin to not rely on stack-heap-promotion
PRs #2745, #2755, #2855 disabled the behavior of heap-promoting stack
variables that were used in non-blocking on blocks.  That meant that this
test was using a stack variable that had gone out of scope. I adjust the
test (by moving a sleep) so that the stack frame with the variable does
not go out of scope before the variable is used. Note that the test
already had such a sleep - I just moved it to the inner scope instead of
at the end of main().