Implement the FileSystem function moveDir
moveDir takes in a source directory and a destination and moves the contents of
the source directory into the destination.  I chose to implement this as a
call to copyTree followed by a call to rmTree on src.

Limitations/caveats: If destination already exists and is a directory we
currently error out, although this is not the final intended behavior - the
implementation to handle this case relies on a couple of Path's string functions
which haven't been implemented yet.  This implementation requires rmTree to be
implemented, so it'll wait until I do that (I'd forgotten the order I wanted to
implement things in), but while I was thinking about it, I figured I'd finish
sketching out the function.

No tests yet.