Handle complex primitives prim_get_real and prim_get_imag for wide complexes
If the argument to prim_get_real or prim_get_imag in a move is wide, the lhs
of the move should be wide also. Add these primitives to the list where this
happens in insertWideReferences.

The runtime functions that implement these primitives take a local address
of a complex and return the local address of the real or imaginary component
by doing an address manipulation.  If the argument to the primitive is wide,
pass the address portion to the runtime function treating it as a local
pointer, then assign the resulting pointer back into a wide reference to the
same locale as the argument.

Update one test .good file that got a more accurate result by using C99
complexes vs. record-based complexes.

Passed full local linux64 and comm-gasnet linux64 testing.