Add a test of rmTree for a large tree structure.
The precomp for this test makes the empty directories that git ignores in the
keeper directory.  Then the test copies the directory we want to keep into a
removable directory (if it did not already exist).  It then checks to make sure
the directory exists, removes it, then checks that the directory is gone.

This commit also adds the directory we want to keep so it doesn't get
regenerated all the time.