Filing a bug against extern types in generic functions
It turns out that if we declare extern types in generic functions,
even those that are instantiated only a single time as in this
example, we hit an internal error.  From what I could tell and guess,
the cause is that we clone everything in the function, including the
extern type, but that since extern types are primitives, and
primitives throw an internal error when they are duplicated, that
causes a problem.  Workarounds include giving the type of the formal
(so the function is no longer generic) or moving the extern type
elsewhere (module level or in its own module).

It seems as though a fix for this _may_ be to not clone extern types
and just return the original type given that there's nothing about
them that's going to need to be specialized from one generic
instantiation to the other.  I'm not sure whether this would best be
done low in the callchain or higher up.  Worth a shot.