Fix bug in handling symlinks to directories.
Since isDir/isFile operate on the link source rather than the link itself, if
a link was to a directory then rmTreeHelper would ignore it.  I wrote a test
ensuring that such links were removed without harming the directory that they
linked to, and fix the implementation of rmTreeHelper to no longer ignore
links and to respond appropriately when they were encountered.