Avoid a copy when returning the result of a call
For code like this:

  proc return2() {
    return return1();
  }

the compiler was adding a temporary to store the result of return1()
and an autoCopy/autoDestroy pair. Since the return value of a function
is the caller's responsibility to free, we can just directly move
a the result from one call into the return value variable for another.

This patch adds a flag, FLAG_RET_EXPR_TEMP, that is set in normalize for
temporaries created in this situation (storing the result of a call that
is then returned). It modifies postFold in functionResolution to avoid
adding FLAG_INSERT_AUTO_COPY and FLAG_INSERT_AUTO_DESTROY for such
variables - except that we leave the old behavior for record wrapped
types, since there seems to be some kind of issue with enabling this
change for them (Hello World seems to enter an infinite loop).

The return-return.chpl test is useful for investigating this case.