Add check that operation and relevant DefExpr are in same block
The test test/types/records/ferguson/variables/local-assign-if.chpl
shows a case where this optimization was previously incorrect. Remedy
by only applying the optimization if the assignment we are optimizing
is in the same block as the DefExpr for the RHS. Otherwise, we would
remove the FLAG_INSERT_AUTO_DESTROY from the RHS variable, but possibly
not PRIM_MOVE into it if the relevant block we are in was not run.