Taking a different approach to this
Instead of "hey, it's an external type so it's OK to share this
primitive TypeSybmol" instead I'm saying "it's an external type,
so it's OK to make a copy of this primitive Type."  This gets
around a reluctance that Mike had about a copy method not actually
making a copy while also potentially making this code more robust
if there were cases where we actually needed to make clones of
external type symbols based on the instantiation of the generic.

It also, in retrospect, seems a bit logical because the symbol
is local to the generic procedure and therefore each instantiation
of the procedure should get its own local copy of that symbol
rather than sharing someone else's which may not even be in
scope.